refactor(ratio-ui): split types.ts into spacing, borders, colors modules
Replace the generic types.ts and classBuilders.ts with domain-specific
files: spacing.ts, borders.ts, colors.ts. Each module co-locates its
types and class builders. Update all imports accordingly.
