+++ libs/ratio-ui/src/tokens/index.ts
@@ -1,13 +1,6 @@
 export { gridGapClasses, getGridClasses, type GapSize } from './spacing';
 
-// ADR-0001: unified spacing, border, and color types
-export type {
-  Space,
-  SpacingProps,
-  BorderVariant,
-  BorderProps,
-  Color,
-  Status,
-} from './types';
-
-export { buildSpacingClasses, buildBorderClasses } from './classBuilders';
+// ADR-0001: unified spacing, border, and color APIs
+export { type Space, type SpacingProps, buildSpacingClasses } from './spacing';
+export { type BorderVariant, type BorderProps, buildBorderClasses } from './borders';
+export { type Color, type Status } from './colors';
